ORC-1224: Move getDecompressor to HadoopShimsCurrent
