+++ java/shims/src/java/org/apache/orc/impl/HadoopShimsPre2_6.java
@@ -20,8 +20,6 @@ package org.apache.orc.impl;
 
 import org.apache.hadoop.conf.Configuration;
 import org.apache.hadoop.fs.FSDataInputStream;
-import org.apache.hadoop.io.compress.snappy.SnappyDecompressor.SnappyDirectDecompressor;
-import org.apache.hadoop.io.compress.zlib.ZlibDecompressor;
 
 import java.io.IOException;
 import java.io.OutputStream;
@@ -38,26 +36,9 @@ import java.util.Random;
  */
 public class HadoopShimsPre2_6 implements HadoopShims {
 
-  static DirectDecompressor getDecompressor( DirectCompressionType codec) {
-    switch (codec) {
-      case ZLIB:
-        return new ZlibDirectDecompressWrapper
-            (new ZlibDecompressor.ZlibDirectDecompressor());
-      case ZLIB_NOHEADER:
-        return new ZlibDirectDecompressWrapper
-            (new ZlibDecompressor.ZlibDirectDecompressor
-                (ZlibDecompressor.CompressionHeader.NO_HEADER, 0));
-      case SNAPPY:
-        return new SnappyDirectDecompressWrapper
-            (new SnappyDirectDecompressor());
-      default:
-        return null;
-    }
-  }
-
   @Override
   public DirectDecompressor getDirectDecompressor(DirectCompressionType codec) {
-    return getDecompressor(codec);
+    return HadoopShimsCurrent.getDecompressor(codec);
  }
 
   @Override
